+++ components/sidebar/sidebar.module.ts
@@ -1,6 +1,8 @@
 import {NgModule} from '@angular/core';
 import {CommonModule} from "@angular/common";
-import {SuiSidebar, SuiSidebarContainer, SuiSidebarSibling} from './sidebar';
+import {SuiSidebar} from './sidebar';
+import {SuiSidebarContainer} from './sidebar-container';
+import {SuiSidebarSibling} from './sidebar-sibling';
 
 @NgModule({
     imports: [
