+++ demo/src/app/app.component.ts
@@ -5,18 +5,4 @@ import {Component, HostBinding} from '@angular/core';
     templateUrl: './app.component.html',
     styleUrls: ['./app.component.css']
 })
-export class AppComponent {
-    @HostBinding('class.pushable')
-    private _pushableClass:boolean = true;
-
-    public menuOpen:boolean = false;
-
-    public openMenu(event:MouseEvent) {
-        event.stopPropagation();
-        this.menuOpen = true;
-    }
-
-    public closeMenu() {
-        this.menuOpen = false;
-    }
-}
+export class AppComponent {}
